Rollback newrelic agent
Chnage 85e5b04 updated the new relic agent to ver 3.39.1. However
in order to deploy the new agent the heroku JAVA_OPTS Env Var pointing
to the agent jar needs to be updated at the same time. There really isn't
a good way to deploy and update the Env var without incurring downtime
on heroku.  AWS provides a rolling update feature which will make this
possible.  We rollback the agent now and update after we move to
using AWS.